LayoutTests/fast/canvas/webgl/array-set-with-offset-expected.txt

Test setting WebGL array with offset

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Regression test for https://bugs.webkit.org/show_bug.cgi?id=76040 : Int16Array.set(array, offset) fails on first execution
Testing Int8Array
PASS webGLArray[0] is 0
PASS webGLArray[1] is 1
PASS webGLArray[2] is 2
PASS webGLArray[3] is bounds[0]
PASS webGLArray[4] is bounds[1]
Testing Uint8Array
PASS webGLArray[0] is 0
PASS webGLArray[1] is 1
PASS webGLArray[2] is 2
PASS webGLArray[3] is bounds[0]
PASS webGLArray[4] is bounds[1]
Testing Int16Array
PASS webGLArray[0] is 0
PASS webGLArray[1] is 1
PASS webGLArray[2] is 2
PASS webGLArray[3] is bounds[0]
PASS webGLArray[4] is bounds[1]
Testing Uint16Array
PASS webGLArray[0] is 0
PASS webGLArray[1] is 1
PASS webGLArray[2] is 2
PASS webGLArray[3] is bounds[0]
PASS webGLArray[4] is bounds[1]
Testing Int32Array
PASS webGLArray[0] is 0
PASS webGLArray[1] is 1
PASS webGLArray[2] is 2
PASS webGLArray[3] is bounds[0]
PASS webGLArray[4] is bounds[1]
Testing Uint32Array
PASS webGLArray[0] is 0
PASS webGLArray[1] is 1
PASS webGLArray[2] is 2
PASS webGLArray[3] is bounds[0]
PASS webGLArray[4] is bounds[1]
Testing Float32Array
PASS webGLArray[0] is 0
PASS webGLArray[1] is 1
PASS webGLArray[2] is 2
PASS webGLArray[3] is bounds[0]
PASS webGLArray[4] is bounds[1]
PASS successfullyParsed is true

TEST COMPLETE
